jnj_audit_copilot/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Sai Vignan M"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langgraph = "0.2.16"
langsmith = "0.1.106"
langchain = "0.2.16"
langchain-openai = "0.1.7"
jupyter = "^1.1.1"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
networkx = "^3.3"
pydantic-settings = "^2.5.2"
python-pptx = "^1.0.2"
langchain-chroma = "^0.1.4"
langchain-community = "0.2.16"
pypdf = "^5.0.0"
unstructured = "^0.15.13"
python-dotenv = "^1.0.1"
python-docx = "1.1.2"
redis = "^5.1.1"
litellm = "^1.51.3"
retry = "^0.9.2"
ipykernel = "^6.29.5"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.8.0"
flake8 = "^7.1.1"
autopep8 = "^2.3.1"

[tool.isort]
profile = "black"

[tool.black]
line-length = 125
target-version = ['py310']

[tool.flake8]
max-line-length = 125
ignore = ["E203", "W503"]
exclude = ".venv, .git, __pycache__, build, dist"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

